python/sql2xlsx/requirements.txt
python-dotenv
python-dotenv[cli]
pymysql
mysql-connector-python
pugsql
openpyxl
jinja2
flask
Flask-MySQLdb
pyyaml
